transfer_listener/api/v1/transfers/serializers.py
from rest_framework import serializers

from transfer_listener.models import Transfer


class TransferSerializer(serializers.ModelSerializer):
    class Meta:
        model = Transfer
        fields = '__all__'
