requirements.txt
Django==5.1.2
web3==7.4.0
djangorestframework==3.15.2
django-filter==24.3
python-dotenv==1.0.1
